% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dice_positions}
\alias{create_dice_positions}
\title{Get Dice Dot Positions as Text Grid}
\usage{
create_dice_positions(n_dots)
}
\arguments{
\item{n_dots}{Integer between 1 and 6}
}
\value{
Character string representing dice dot layout
}
\description{
Returns a string representing dice layout with numbered positions.
}
\keyword{internal}
