% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_datasets.R
\docType{data}
\name{sample_dice_data2}
\alias{sample_dice_data2}
\title{Example dice dataset 2}
\format{
A data frame; similar structure to \code{sample_dice_data1}.
}
\usage{
data(sample_dice_data2)
}
\description{
A second toy dataset for \code{geom_dice()} examples.
}
\keyword{datasets}
