% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_dots_discrete}
\alias{scale_dots_discrete}
\title{Discrete Scale for Dice Dot Colors}
\usage{
scale_dots_discrete(..., aesthetics = "dots")
}
\arguments{
\item{...}{Passed to \code{ggplot2::discrete_scale()}}

\item{aesthetics}{Character string of the target aesthetic (default: "dots")}
}
\value{
A ggplot2 scale
}
\description{
Creates a ggplot2 discrete scale for dice dot aesthetics.
}
