% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_dice}
\alias{theme_dice}
\title{Dice Theme for ggplot2}
\usage{
theme_dice(x_length, y_length, ...)
}
\arguments{
\item{x_length}{Width of the plotting area}

\item{y_length}{Height of the plotting area}

\item{...}{Additional arguments passed to \code{theme_grey()}}
}
\value{
A ggplot2 theme
}
\description{
A minimal ggplot2 theme for dice plots with controlled aspect ratio.
}
