% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_datasets.R
\docType{data}
\name{sample_dice_large}
\alias{sample_dice_large}
\title{Larger example dice dataset}
\format{
A data frame; similar structure to \code{sample_dice_data1}.
}
\usage{
data(sample_dice_large)
}
\description{
Larger illustrative dataset for more complex dice-plots.
}
\keyword{datasets}
